+++ test/integration/test_brute_force.py
@@ -253,3 +253,30 @@ class TestBruteForce(TestCaseIntegration):
         expected = self.data_dir/'fuzz/23/brute_force.txt'
         reference = self.data_dir/'downsampled_reference/ENST00000265138.4-ENST00000650150.1'
         self.default_test_case(gvf, reference, expected)
+
+    def test_brute_force_fuzz_24(self):
+        """ Fuzz test 24 """
+        gvf = [
+            self.data_dir/'fuzz/24/fake_variants.gvf'
+        ]
+        expected = self.data_dir/'fuzz/24/brute_force.txt'
+        reference = self.data_dir/'downsampled_reference/ENST00000265138.4-ENST00000650150.1'
+        self.default_test_case(gvf, reference, expected)
+
+    def test_brute_force_fuzz_25(self):
+        """ Fuzz test 25 """
+        gvf = [
+            self.data_dir/'fuzz/25/fake_variants.gvf'
+        ]
+        expected = self.data_dir/'fuzz/25/brute_force.txt'
+        reference = self.data_dir/'downsampled_reference/ENST00000265138.4-ENST00000650150.1'
+        self.default_test_case(gvf, reference, expected)
+
+    def test_brute_force_fuzz_26(self):
+        """ Fuzz test 26 """
+        gvf = [
+            self.data_dir/'fuzz/26/fake_variants.gvf'
+        ]
+        expected = self.data_dir/'fuzz/26/brute_force.txt'
+        reference = self.data_dir/'downsampled_reference/ENST00000265138.4-ENST00000650150.1'
+        self.default_test_case(gvf, reference, expected)